requirements.txt
git+ssh://git@github.com/frankroeder/lanro-gym.git
gym==0.21.0
torch
mpi4py
wandb
hydra-core==1.1.1
opencv-python
